Title: 최솟값 3개, Time: 63ms, Memory: 1MB, Status: Passed - Codetree

+++ 260111/최솟값 3개/top-3-smallest-number.cpp
@@ -23,7 +23,7 @@ int main() {
             continue;
         }
 
-        int num[3] = {};
+        long long num[3] = {};
 
         for (int j = 0; j < 3; j++)
         {
